fix(auth-service): remove duplicate model and service defs, keep single source of truth
