+++ auth-microservice/auth-service/model.go
@@ -1,0 +1,9 @@
+package main
+
+type User struct {
+	ID        int
+	Email     string
+	Password  string
+	Role      string
+	CreatedAt string // หรือ time.Time ถ้า import "time"
+}
